Added customCommands option, which allow definition of new commands
customCommands can be inserted into the options passed to $.wysiwyg() .

It consists of key/value pairs, where the key is the name of the custom command (that can be used for to couple a button in the toolbar using the data-edit attribute). The value is an object containing a function with key "execCommand", which actually executes the command, and an "updateToolbar" function, that returns a boolean as to whether the button should be highlighted or not.

A new example file custom-commands.html was added with an example of how to use this, demonstrating a button that unwraps removes the heading (h1, h2, etc) tag around text, and a button that toggles a numbered list.